requirements.txt
customtkinter==5.2.1
requests==2.31.0
PyGithub==2.1.1
pyinstaller==6.3.0
matplotlib==3.8.2
pandas==2.1.4
pillow==10.1.0